Add some tests for glui.actions
Adds some basic tests for glui.actions and makes it possible for
HeadlessBackend to render frames as SVG files. This is very basic and
omits textures, but is mighty handy for visually debugging the output in
a web browser.

At the same time, this also finds what might be considered a bug in
GluiScrollable :D