Minor improvements and AMD bugfix
Please don't bother asking me why or how. If you resized the showcase
window while viewing any article (not the homepage), it would hang.
glGetError appears to fix it, but maybe the issue is still present in
different conditions. Maybe it's not. RaylibBackend is not affected. I
asked in like three different places. Not a single person has a clue.

Works on my machine, also works on my roommate's, so hopefully it works
on yours too.