Don't draw text to irrelevant chunks
If every piece of text would be discarded when drawing to a chunk, the
chunk is skipped. This way, text rendering remains fairly smooth even
when there's tons of it.

Slow down is still present. Ideally, TextRuler should be cached on chunk
boundaries, but this is out of scope for this merge.