Implement a code editor
The editor comes with syntax highlighting and intelligent indent
functionality. This editor is moved from another repository, so this
is currently disabled as it needs to be adapted to fit Fluid. A
subpackage will be made available that will provide integration with
TreeSitter

The current version is probably good enough as a base code editor.
Syntax highlighting and indenting can be provided by the programmer.

Note that the current implementation does not handle automatic
outdenting well; a modification to TextInput is necessary to correctly
detect changes in indentation while typing. Furthermore, indents are
completely opinionated as they're absolute by default. All of these
will be corrected in subsequent commits of this branch.