Improvements to input actions
A failed attempt to make better use of compiler caching to speed
up input action generation.

This doesn't appear to improve the compile-time performance of
input handling a lot , but at least it does make the code more
readable.

Sadly, function bodies inside mixin templates also contribute
to compile times in dependencies, even if they are completely
unnecessary (already compiled). This seems to be a flaw in the
compiler, so perhaps it can be fixed in DMD itself.