Prepare DragSlot for new I/O
There is currently no way for a DragAction child to be resized while
drawing. Resizing in DragAction would mean changing the set of I/O
systems available to the child, to only the ones that were made
available to the `HoverIO` controlling the pointer. As a consequence,
some systems would be discarded, including the `HoverIO` system
itself. Even if the node was moved as a mocked child of `HoverIO`,
the set of I/O systems applied by the node may vary depending on the
node's conditions which cannot be simulated by the action.

Consider a situation like:

    auto frame1 = implementIO!HoverIO();
    auto frame2 = implementIO!FocusIO();
    drawChild(child1);
    frame2.stop();
    super.resizeImpl(space);
    frame1.stop();

Depending on where, during this resize, the node is inserted, its
I/O system list may or may not include the `FocusIO` system.
Similarly, a node may opt not to implement any systems, based on
a condition such as the presence of a child:

     if (hasChildren) {
         auto frame = this.implementIO();
         super.resizeImpl(space);
     }

Thus, the only reasonable solution out of this, is to lock a node's
set of systems, preventing it from making changes. This could mean
turning `use` and `require` into NO-OPs, aside from having `require`
test if the requested system is present, keeping in mind that the
drag slot's children can also implement I/O systems, and a system
swap or a newly added node may be one of the reasons why such a resize
is needed in the first place.